+++ packages/docs/.vitepress/theme/examples/menu/index.vue
@@ -1,9 +1,0 @@
-<script setup>
-import Base from "./base.vue";
-import BaseCode from "./base.vue?raw";
-</script>
-
-<template>
-    <h3 id="base">Base</h3>
-    <ExampleViewer :component="Base" :code="BaseCode" />
-</template>