+++ packages/oruga/src/components/datepicker/examples/range.vue
@@ -1,0 +1,20 @@
+<script setup lang="ts">
+import { ref } from "vue";
+
+const date = ref();
+</script>
+
+<template>
+    <section>
+        <o-field label="Select a date">
+            <o-datepicker
+                v-model="date"
+                placeholder="Click to select..."
+                icon="calendar"
+                trap-focus
+                range />
+        </o-field>
+
+        <p><b>Selected:</b> {{ date }}</p>
+    </section>
+</template>