+++ packages/oruga/src/components/dropdown/examples/inline.vue
@@ -1,0 +1,13 @@
+<template>
+    <section class="odocs-spaced">
+        <o-dropdown inline>
+            <template #trigger>
+                <o-button label="Inline" />
+            </template>
+
+            <o-dropdown-item :value="1" label="Action " />
+            <o-dropdown-item :value="2" label="Another action " />
+            <o-dropdown-item :value="3" label="Something else " />
+        </o-dropdown>
+    </section>
+</template>