+++ packages/oruga/src/components/dropdown/tests/dropdownitem.test.ts
@@ -3,7 +3,7 @@ import { mount } from "@vue/test-utils";
 
 import ODropdownItem from "@/components/dropdown/DropdownItem.vue";
 
-describe("ODropdownItem", () => {
+describe("ODropdownItem tests", () => {
     test("expect should wrap in dropdown", () => {
         expect(() => mount(ODropdownItem)).toThrowError("should wrap");
     });