+++ packages/oruga/src/components/radio/Radio.vue
@@ -2,12 +2,7 @@
 import { computed, ref, type PropType } from "vue";
 
 import { getOption } from "@/utils/config";
-import { uuid } from "@/utils/helpers";
-import {
-    defineClasses,
-    useVModelBinding,
-    useInputHandler,
-} from "@/composables";
+import { defineClasses, useInputHandler } from "@/composables";
 
 import type { ComponentClass } from "@/types";
 
@@ -54,8 +49,6 @@ const props = defineProps({
     required: { type: Boolean, default: false },
     /** Same as native name */
     name: { type: String, default: undefined },
-    /** Accessibility label to establish relationship between the checkbox and control label */
-    ariaLabelledby: { type: String, default: () => uuid() },
     /** Same as native autocomplete options to use in HTML5 validation */
     autocomplete: {
         type: String,
@@ -147,7 +140,7 @@ const { onBlur, onFocus, onInvalid, setFocus } = useInputHandler(
     props,
 );
 
-const vmodel = useVModelBinding<string | number | boolean>(props, emits);
+const vmodel = defineModel<string | number | boolean>({ default: undefined });
 
 const isChecked = computed(() => vmodel.value === props.nativeValue);
 
@@ -214,17 +207,13 @@ defineExpose({ focus: setFocus });
             :name="name"
             :autocomplete="autocomplete"
             :value="nativeValue"
-            :aria-labelledby="ariaLabelledby"
             @click.stop
             @blur="onBlur"
             @focus="onFocus"
             @invalid="onInvalid"
             @input="onInput" />
 
-        <span
-            v-if="label || $slots.default"
-            :id="ariaLabelledby"
-            :class="labelClasses">
+        <span v-if="label || $slots.default" :class="labelClasses">
             <!--
                 @slot Override the label, default is label prop 
             -->