+++ CHANGELOG.md
@@ -1,3 +1,94 @@
+## [0.8.11](https://github.com/oruga-ui/oruga/compare/v0.8.10...v0.8.11) (2024-06-05)
+
+
+### Bug Fixes
+
+* **datepicker:** always validate programmatic date updates ([#927](https://github.com/oruga-ui/oruga/issues/927)) ([20ceedd](https://github.com/oruga-ui/oruga/commit/20ceedda2566211b5c790a12a687a01ce26380b5))
+* **datepicker:** update focused date on editing directly ([#941](https://github.com/oruga-ui/oruga/issues/941)) ([1af13cf](https://github.com/oruga-ui/oruga/commit/1af13cf1cc97703aa58820d99b403791b554f9ba))
+* **taginput:** solve defineModel macro array constrains ([#942](https://github.com/oruga-ui/oruga/issues/942)) ([404c5f6](https://github.com/oruga-ui/oruga/commit/404c5f6d1aa61cbbbe7dfce089bf203ffde4259e))
+
+
+### Features
+
+* **inputs:** detect constraint validation attribute changes ([#930](https://github.com/oruga-ui/oruga/issues/930)) ([4de3bd7](https://github.com/oruga-ui/oruga/commit/4de3bd74f512a3a1b07363b0cb062d6f2b5cc95d))
+
+
+
+## [0.8.10](https://github.com/oruga-ui/oruga/compare/v0.8.9...v0.8.10) (2024-06-03)
+
+
+### Bug Fixes
+
+* **datetimpicker:** fix inline value edit | make picker props reactive ([#932](https://github.com/oruga-ui/oruga/issues/932)) ([1e93773](https://github.com/oruga-ui/oruga/commit/1e93773fbbe38859ecac3deee8e651c340d01012))
+
+
+### Features
+
+* **modal:** set default `destroyOnHide` to `true` for programmatic usage ([#937](https://github.com/oruga-ui/oruga/issues/937)) ([0b0bdca](https://github.com/oruga-ui/oruga/commit/0b0bdca61e72bf471d2e8788d8737a47803f79b4))
+* **steps:** add `destroyOnHide` prop ([#934](https://github.com/oruga-ui/oruga/issues/934)) ([d06ea1e](https://github.com/oruga-ui/oruga/commit/d06ea1ecf462b54c626a72286ba123d2a1489031))
+* **tabs:** add `destroyOnHide` prop ([#926](https://github.com/oruga-ui/oruga/issues/926)) ([e87882b](https://github.com/oruga-ui/oruga/commit/e87882bcb633ba288228f77ddf885a3373bbd860))
+
+
+
+## [0.8.9](https://github.com/oruga-ui/oruga/compare/v0.8.8...v0.8.9) (2024-05-01)
+
+
+### Bug Fixes
+
+* **field:** fix bug when prop `horizontal` is set ([#915](https://github.com/oruga-ui/oruga/issues/915)) ([962bfea](https://github.com/oruga-ui/oruga/commit/962bfea1d447aac2b6a2c3174e2becf49ebac926))
+* **table:** fix filtering when table is not sorted ([#911](https://github.com/oruga-ui/oruga/issues/911)) ([10ee451](https://github.com/oruga-ui/oruga/commit/10ee45119d107ae35c9c25fc17f9dfe8d12116c3))
+* **Table:** fix table async example backend sort function ([#910](https://github.com/oruga-ui/oruga/issues/910)) ([18f3c00](https://github.com/oruga-ui/oruga/commit/18f3c002bc295601948d111b005bc3f1e221cba5))
+* **taginput:** fix infinite recursive update when updating modeValue ([#917](https://github.com/oruga-ui/oruga/issues/917)) ([e2e3c4a](https://github.com/oruga-ui/oruga/commit/e2e3c4a2d9347a3e01709379a4c6ee7bb98eba5f))
+
+
+### Features
+
+* **table:** make debounceSearch a global option ([#912](https://github.com/oruga-ui/oruga/issues/912)) ([517598f](https://github.com/oruga-ui/oruga/commit/517598f5220db093c282a108a427a36f97502e7d))
+
+
+
+## [0.8.8](https://github.com/oruga-ui/oruga/compare/v0.8.7...v0.8.8) (2024-04-23)
+
+
+### Bug Fixes
+
+* **build:** fix cjs extension in package.json entry point ([#894](https://github.com/oruga-ui/oruga/issues/894)) ([51e7184](https://github.com/oruga-ui/oruga/commit/51e718411014ec2317d38dae2dacc12cededa4b7))
+* **datepicker:** fix defineClasses parent props problem ([#900](https://github.com/oruga-ui/oruga/issues/900)) ([087a8dc](https://github.com/oruga-ui/oruga/commit/087a8dcf55c59712d7d31c0b14c0e12a9dde589d))
+* **datepicker:** make date parsing work again ([#882](https://github.com/oruga-ui/oruga/issues/882)) ([047e999](https://github.com/oruga-ui/oruga/commit/047e999324b912e8b7b456bb4f72feb09b4282e8))
+* **field:** add field `setInputId` & remove `ariaLabelledby` prop ([#902](https://github.com/oruga-ui/oruga/issues/902)) ([458d45c](https://github.com/oruga-ui/oruga/commit/458d45c17d2c3d96813e78f1902cd781d7e6ae93))
+* **field:** fix slot invoked outside render function & ass `messageTag` prop ([#906](https://github.com/oruga-ui/oruga/issues/906)) ([f9d07c6](https://github.com/oruga-ui/oruga/commit/f9d07c62a091bf44a36b38454c5b1ba86559270a))
+* **menu:** fix `useProviderParent` composable `childItems` type problem ([#897](https://github.com/oruga-ui/oruga/issues/897)) ([b9656a6](https://github.com/oruga-ui/oruga/commit/b9656a6b5adfc27510be66e3bfe9b018e5de2cdd))
+* **taginput:** fix selection with autocomplete on non string items ([#889](https://github.com/oruga-ui/oruga/issues/889)) ([930c397](https://github.com/oruga-ui/oruga/commit/930c39707d2f938fe87876808c90251a743b8cd2))
+
+
+### Features
+
+* **datepicker:** change prop `readonly` default to `false` & disable calendar interaction while `readonly` ([#901](https://github.com/oruga-ui/oruga/issues/901)) ([364d0c5](https://github.com/oruga-ui/oruga/commit/364d0c5c9128e57167cff3922a0585bf6d323010))
+* **taginput:** add `removeItem` as selected slot property ([#887](https://github.com/oruga-ui/oruga/issues/887)) ([#888](https://github.com/oruga-ui/oruga/issues/888)) ([c8c26cc](https://github.com/oruga-ui/oruga/commit/c8c26cc7608a4ecdcc7336f63a0a069721a9cfcc))
+* **taginput:** update prop types ([#895](https://github.com/oruga-ui/oruga/issues/895)) ([b5709c5](https://github.com/oruga-ui/oruga/commit/b5709c569a8390e0f7944234b166534d146a363d))
+
+
+
+## [0.8.7](https://github.com/mlmoravek/oruga/compare/v0.8.6...v0.8.7) (2024-04-02)
+
+
+### Bug Fixes
+
+* **autocomplete:** adjust active state behavior ([#859](https://github.com/mlmoravek/oruga/issues/859)) ([9ce1971](https://github.com/mlmoravek/oruga/commit/9ce1971c71e719a392dbfa7735dc64c5cabaeef2))
+* **autocomplete:** change options type definition to generic ([#866](https://github.com/mlmoravek/oruga/issues/866)) ([5f6717a](https://github.com/mlmoravek/oruga/commit/5f6717a668be58ee5916b55c896034574de02856))
+* **steps|tabs:** fix disable animation ([#870](https://github.com/mlmoravek/oruga/issues/870)) ([08c1fbf](https://github.com/mlmoravek/oruga/commit/08c1fbf93adda4af411e733a75fed097432d5da7))
+* **steps|tabs:** make item props reactive again ([#879](https://github.com/mlmoravek/oruga/issues/879)) ([269183d](https://github.com/mlmoravek/oruga/commit/269183dcadb918b365894cef4809a12241387172))
+* **table:** make `TableColumn` props reactive again ([#878](https://github.com/mlmoravek/oruga/issues/878)) ([d2b1caa](https://github.com/mlmoravek/oruga/commit/d2b1caa3074b7f3826e856942e7979b9abc72aea))
+* fix type generation script ([#867](https://github.com/mlmoravek/oruga/issues/867)) ([2f6b6c8](https://github.com/mlmoravek/oruga/commit/2f6b6c893b75997e95be1e276616dd750f5c99a1))
+* resolve maximum recursive update exceeded in tests ([#869](https://github.com/mlmoravek/oruga/issues/869)) ([40d8b94](https://github.com/mlmoravek/oruga/commit/40d8b942805e15d07589e9bc93b6841885f5cab3))
+
+
+### Features
+
+* **tabs:** move inherit attrs to div instead of transition ([#871](https://github.com/mlmoravek/oruga/issues/871)) ([6db3f74](https://github.com/mlmoravek/oruga/commit/6db3f74c0a24b13d4ce7f9edccaa131829640131))
+
+
+
 ## [0.8.6](https://github.com/oruga-ui/oruga/compare/v0.8.5...v0.8.6) (2024-03-11)